% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoqmsp-package.R
\docType{package}
\name{autoqmsp-package}
\alias{autoqmsp}
\alias{autoqmsp-package}
\title{autoqmsp: automated qMSP analysis}
\description{
Typical workflow:
```
runs <- read_eds_files(c("run1.eds", "run2.eds"))
res  <- analyze_qmsp(runs, reference = "B ACTIN")
res$results
plot_methylation(res)
export_results(res, "results.xlsx")
```
Or start the point-and-click app with [run_app()].
}
\author{
\strong{Maintainer}: Nazli Akilli \email{hikmet.akilli@gmail.com}

}
\keyword{internal}
