% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoqmsp-package.R
\name{run_app}
\alias{run_app}
\title{Start the autoqmsp app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to [shiny::runApp()].}
}
\description{
Opens a browser app where you upload `.eds` files, adjust settings and
download an Excel report. Needs the `shiny` package.
}
